+++ src/parser.rs
@@ -2,6 +2,7 @@ use crate::linereader::LineReader;
 use crate::ParseError;
 use crate::Section;
 use std::io;
+use std::path::Path;
 
 /// Parser for the text of an EditorConfig file.
 ///
@@ -13,6 +14,8 @@ pub struct ConfigParser<R: io::BufRead> {
     pub is_root: bool,
     eof: bool,
     reader: LineReader<R>,
+    #[cfg(feature = "track-source")]
+    path: Option<std::sync::Arc<Path>>,
 }
 
 impl<R: io::Read> ConfigParser<io::BufReader<R>> {
@@ -22,14 +25,33 @@ impl<R: io::Read> ConfigParser<io::BufReader<R>> {
     pub fn new_buffered(source: R) -> Result<ConfigParser<io::BufReader<R>>, ParseError> {
         Self::new(io::BufReader::new(source))
     }
+    /// Convenience function for construction using an unbuffered [`io::Read`]
+    /// which is assumed to be a file at `path`.
+    ///
+    /// See [`ConfigParser::new_with_path`].
+    pub fn new_buffered_with_path(
+        source: R,
+        path: Option<impl Into<std::sync::Arc<Path>>>,
+    ) -> Result<ConfigParser<io::BufReader<R>>, ParseError> {
+        Self::new_with_path(io::BufReader::new(source), path)
+    }
 }
 
 impl<R: io::BufRead> ConfigParser<R> {
-    /// Constructs a new [`ConfigParser`] and reads the preamble from the provided source.
+    /// Constructs a new [`ConfigParser`] and reads the preamble from the provided source,
+    /// which is assumed to be a file at `path`.
     ///
     /// Returns `Ok` if the preamble was parsed successfully,
     /// otherwise returns `Err` with the error that occurred during reading.
-    pub fn new(buf_source: R) -> Result<ConfigParser<R>, ParseError> {
+    ///
+    /// If the `track-source` feature is enabled and `path` is `Some`,
+    /// [`RawValue`][crate::rawvalue::RawValue]s produced by this parser will
+    /// have their sources set appropriately.
+    /// Otherwise, `path` is unused.
+    pub fn new_with_path(
+        buf_source: R,
+        #[allow(unused)] path: Option<impl Into<std::sync::Arc<Path>>>,
+    ) -> Result<ConfigParser<R>, ParseError> {
         let mut reader = LineReader::new(buf_source);
         let mut is_root = false;
         let eof = loop {
@@ -49,11 +71,22 @@ impl<R: io::BufRead> ConfigParser<R> {
                 }
             }
         };
+        #[cfg(feature = "track-source")]
+        let path = path.map(Into::into);
         Ok(ConfigParser {
             is_root,
             eof,
             reader,
+            #[cfg(feature = "track-source")]
+            path,
         })
+    }
+    /// Constructs a new [`ConfigParser`] and reads the preamble from the provided source.
+    ///
+    /// Returns `Ok` if the preamble was parsed successfully,
+    /// otherwise returns `Err` with the error that occurred during reading.
+    pub fn new(buf_source: R) -> Result<ConfigParser<R>, ParseError> {
+        Self::new_with_path(buf_source, Option::<std::sync::Arc<Path>>::None)
     }
 
     /// Returns `true` if there may be another section to read.
@@ -75,6 +108,9 @@ impl<R: io::BufRead> ConfigParser<R> {
         if let Ok(Line::Section(header)) = self.reader.reparse() {
             let mut section = Section::new(header);
             loop {
+                // Get line_no here to avoid borrowing issues, increment for 1-based indices.
+                #[cfg(feature = "track-source")]
+                let line_no = self.reader.line_no() + 1;
                 match self.reader.next_line() {
                     Err(e) => {
                         self.eof = true;
@@ -87,7 +123,13 @@ impl<R: io::BufRead> ConfigParser<R> {
                     Ok(Line::Section(_)) => break Ok(section),
                     Ok(Line::Nothing) => (),
                     Ok(Line::Pair(k, v)) => {
-                        section.insert(k, v.to_owned());
+                        #[allow(unused_mut)]
+                        let mut v = crate::rawvalue::RawValue::from(v.to_owned());
+                        #[cfg(feature = "track-source")]
+                        if let Some(path) = self.path.as_ref() {
+                            v.set_source(path.clone(), line_no);
+                        }
+                        section.insert(k, v);
                     }
                 }
             }